+++ packages/deployment-service/src/executables/deploy-react.ts
@@ -1,4 +1,5 @@
-import { PipelineEntity, TaskEntity } from '@/entities'
+import { PipelineEntity } from './../entities/pipeline.entity'
+import { TaskEntity } from './../entities/task.entity'
 import { ExecutableType } from './executable'
 
 export const deployReact: ExecutableType = (task: TaskEntity, pipeline: PipelineEntity): Promise<boolean> => {
