+++ packages/deployment-service/src/executables/pull.ts
@@ -1,4 +1,5 @@
-import { PipelineEntity, TaskEntity } from '@/entities'
+import { PipelineEntity } from './../entities/pipeline.entity'
+import { TaskEntity } from './../entities/task.entity'
 import { ExecutableType } from './executable'
 
 export const pull: ExecutableType = (task: TaskEntity, pipeline: PipelineEntity): Promise<boolean> => {
